An attempt to draw connecting lines between jobs and the scoring lines.
An example job configuration would look like

    [Options]
    # ...
    ScoringStyle = surround
    ScoringLineLayers = *toplayer
    ScoringLineWidth  = 1

    [myjob]
    # ...
    # Points to draw a connecting line to the scoring
    #  TopX,TopY,  RightX,RightY,  BottomX,BottomY,  LeftX,LeftY
    ScoringLineConnectionPoints = 19,54.8, 29.5,29.5, 23,10.473,  10.44,34.0
    ScoringLineConnectionLayer  = *toplayer

This doesn't work,

  * I couldn't get rotation working
  * the polyline for the connection "trace" would need square ends, I'd need to look into gerber more than I care to to figure that out
  * it's probably wasted effort anyway

The only reason for doing this is to try and work around board-houses price increases on DIY panel when they are not actually scored/tab-routed anyway.  Houses like AllPCB and PCBWay charge a lot (LOT) more if you have boards on a diy panel like that.

The idea was to make the indivdual jobs all "connected" by specifying a suitable connection point on each side, and then a "tributary trace" would be drawn from one of those points per job feeding into the nearest ScoringLine (by only connecting one of the points we avoid accidental short circuits).  So it would look like every job was connected electrically to every other job via the "scoring lines", all of this drawn in a copper layer.

This is probably wasted effort because I have no idea if this would work to "fool" the board houses, and even if it did, they might just as easily decide to not accept it if they caught on (and it wouldn't exactly be hard to see).